MyJobs (Failures 2, Errors 3): Ignore migrations on Analytics during tests.
django-1.6 was not running migrations during tests, so many of our tests create
data that was normally created by data migrations.  This causes errors now that
migrations are run when django executes tests.  So we tell it to ignore them.
In future, we should change our code to execute data migrations during tests.